MedAIClerk-Backend/requirements.txt
accelerate>=0.31.0
fastapi==0.111.0
fastapi>=0.111.0
openai==1.54.1
openai>=1.30.0
pydantic==2.7.1
pydantic>=2.7.1
python-dotenv==1.0.1
python-dotenv>=1.0.1
sentencepiece>=0.2.0
torch>=2.3.0
transformers>=4.42.0
uvicorn[standard]==0.30.0
uvicorn[standard]>=0.30.0

# bitsandbytes is installed in CUDA image (Dockerfile.cuda)
